Move App.tsx inside components
